UIE-190 Auth Fetch Breakout part-4  - finished move of authOpts and withAuthSession references to now come from auth-session.ts module.  - removed pass-thru exports for these in auth-fetch.ts and switched over import and mock references as needed.
