+++ src/libs/ajax/Metrics.ts
@@ -2,7 +2,7 @@ import { getDefaultProperties } from '@databiosphere/bard-client';
 import { jsonBody } from '@terra-ui-packages/data-client-core';
 import _ from 'lodash/fp';
 import { ensureAuthSettled } from 'src/auth/auth';
-import { authOpts } from 'src/auth/auth-fetch';
+import { authOpts } from 'src/auth/auth-session';
 import { fetchBard } from 'src/libs/ajax/fetch/fetchBard';
 import { getConfig } from 'src/libs/config';
 import { withErrorIgnoring } from 'src/libs/error';
