+++ src/libs/ajax/TermsOfService.ts
@@ -1,5 +1,5 @@
 import _ from 'lodash/fp';
-import { authOpts } from 'src/auth/auth-fetch';
+import { authOpts } from 'src/auth/auth-session';
 import { fetchSam } from 'src/libs/ajax/ajax-common';
 
 export interface SamUserTermsOfServiceDetails {
