+++ src/libs/ajax/leonardo/Disks.test.ts
@@ -1,6 +1,6 @@
 import { FetchFn } from '@terra-ui-packages/data-client-core';
 import { azureDisk, galaxyDisk, undecoratePd } from 'src/analysis/_testData/testData';
-import { authOpts } from 'src/auth/auth-fetch';
+import { authOpts } from 'src/auth/auth-session';
 import { fetchLeo } from 'src/libs/ajax/ajax-common';
 import { Disks } from 'src/libs/ajax/leonardo/Disks';
 import { asMockedFn } from 'src/testing/test-utils';
@@ -28,11 +28,11 @@ jest.mock('src/libs/ajax/ajax-common', (): Partial<AjaxCommonExports> => {
   return mocks;
 });
 
-type AuthFetchExports = typeof import('src/auth/auth-fetch');
-jest.mock('src/auth/auth-fetch', (): AuthFetchExports => {
+type AuthSessionExports = typeof import('src/auth/auth-session');
+jest.mock('src/auth/auth-session', (): AuthSessionExports => {
   const { asMockedFn } = jest.requireActual<TestUtilsExports>('src/testing/test-utils');
-  const mocks: AuthFetchExports = {
-    ...jest.requireActual<AuthFetchExports>('src/auth/auth-fetch'),
+  const mocks: AuthSessionExports = {
+    ...jest.requireActual<AuthSessionExports>('src/auth/auth-session'),
     authOpts: jest.fn(),
     withAuthSession: jest.fn(),
   };
