+++ src/libs/ajax/workflows-app/Cbas.js
@@ -1,7 +1,7 @@
 import { jsonBody } from '@terra-ui-packages/data-client-core';
 import _ from 'lodash/fp';
 import qs from 'qs';
-import { authOpts } from 'src/auth/auth-fetch';
+import { authOpts } from 'src/auth/auth-session';
 import { fetchFromProxy } from 'src/libs/ajax/ajax-common';
 
 export const Cbas = (signal) => ({
